LayoutTests/storage/indexeddb/modern/idbdatabase-deleteobjectstore-failures-expected.txt

ALERT: Initial upgrade needed: Old version - 0 New version - 1
ALERT: Initial upgrade versionchange transaction complete
ALERT: readwrite put success - about to try to delete an objectstore
ALERT: Failed to deleteObjectStore without a versionchange transaction - Error: InvalidStateError: DOM IDBDatabase Exception 11
ALERT: readwrite transaction complete
ALERT: Second upgrade needed: Old version - 1 New version - 2
ALERT: Failed to deleteObjectStore with a non-existent objectstore - Error: NotFoundError: DOM IDBDatabase Exception 8
ALERT: Failed to deleteObjectStore with an in-progress versionchange transaction that is inactive - Error: TransactionInactiveError: DOM IDBDatabase Exception 0
ALERT: Second version change transaction complete
ALERT: Done
This tests some obvious failures that can happen while calling IDBDatabase.deleteObjectStore()
